// src/schemas.ts
export const metricAlertSchema = z.object({
    action: z.string(),
    actor: z.object({ id: z.string(), name: z.string(), type: z.string() }),
    data: z.object({
        description_text: z.string(),
        description_title: z.string(),
        metric_alert: z
            .object({
                alert_rule: z
                    .object({
                        aggregate: z.string(),
                        created_by: z.string().nullish(),
                        dataset: z.string(),
                        date_created: z.string(),
                        date_modified: z.string(),
                        environment: z.string().nullish(),
                        id: z.string(),
                        include_all_projects: z.boolean(),
                        name: z.string(),
                        organization_id: z.string(),
                        projects: z.array(z.string()),
                        query: z.string(),
                        resolution: z.number(),
                        resolve_threshold: z.unknown(),
                        status: z.number(),
                        threshold_period: z.number(),
                        threshold_type: z.number(),
                        time_window: z.number(),
                        triggers: z.array(z.unknown()),
                    })
                    .partial()
                    .nullish(),
                date_closed: z.coerce.date().nullish(),
                date_created: z.coerce.date().nullish(),
                date_detected: z.coerce.date().nullish(),
                date_started: z.coerce.date().nullish(),
                id: z.string(),
                identifier: z.string().nullish(),
                organization_id: z.string(),
                projects: z.array(z.string()),
                status: z.number().nullish(),
                status_method: z.number().nullish(),
                title: z.string(),
                type: z.number(),
            })
            .partial(),
        web_url: z.string(),
    }),
    installation: z.object({ uuid: z.string() }),
});